clean up lib
